+++ src/utils.h
@@ -69,7 +69,7 @@ int addr_equal6(const struct in6_addr *addr1, const struct in6_addr *addr2);
 
 const char *format_mac(char buf[18], const struct mac *addr);
 const char *format_duration(char buf[64], time_t from, time_t to);
-const char *format_size(char buf[64], unsigned bytes);
+const char *format_size(char buf[64], uint64_t bytes);
 
 uint32_t in6_addr_id(const struct in6_addr *addr);
 uint32_t in4_addr_id(const struct in_addr *addr);
